Update: Dataset to hold larger values

+++ dataset_generator.cpp
@@ -8,11 +8,11 @@
 using namespace std;
 
 int main(int argc, char* argv[]) {
-    int num_rows = 1000;
+    long long num_rows = 1000;
     string filename = "";
     
     if (argc > 1) {
-        num_rows = stoi(argv[1]);
+        num_rows = stoll(argv[1]);
     }
     if (argc > 2) {
         filename = argv[2];
@@ -20,8 +20,6 @@ int main(int argc, char* argv[]) {
         filename = "dataset_" + to_string(num_rows) + ".csv";
     }
     
-    // Create seed from "243UC247D5"
-    // Convert to numeric seed by hashing or using as-is
     unsigned long seed = 0;
     string seed_str = "243UC247D5";
     for (char c : seed_str) {
@@ -32,10 +30,10 @@ int main(int argc, char* argv[]) {
     cout << "Using seed: " << seed << endl;
     
     mt19937 generator(seed);
-    uniform_int_distribution<long long> int_dist(1000000000LL, 9999999999LL); // 10-digit numbers
-    uniform_int_distribution<int> char_dist(0, 25); // a-z
+    uniform_int_distribution<long long> int_dist(1000000000LL, 9999999999LL); 
+    uniform_int_distribution<int> char_dist(0, 25); 
     
-    set<long long> used_numbers; // Track unique numbers
+    set<long long> used_numbers; 
     
     ofstream file(filename);
     if (!file.is_open()) {
@@ -43,17 +41,15 @@ int main(int argc, char* argv[]) {
         return 1;
     }
     
-    // Write header
     file << "ID,NAME" << endl;
     
-    int generated = 0;
-    int attempts = 0;
-    int max_attempts = num_rows * 100; // Prevent infinite loop
+    long long generated = 0;
+    long long attempts = 0;
+    long long max_attempts = num_rows * 100; 
     
     while (generated < num_rows && attempts < max_attempts) {
         long long random_id = int_dist(generator);
         
-        // Check uniqueness
         if (used_numbers.find(random_id) != used_numbers.end()) {
             attempts++;
             continue;
@@ -61,13 +57,11 @@ int main(int argc, char* argv[]) {
         
         used_numbers.insert(random_id);
         
-        // Generate 5-letter lowercase string
         string random_str = "";
         for (int i = 0; i < 5; i++) {
             random_str += (char)('a' + char_dist(generator));
         }
         
-        // Write to CSV
         file << random_id << "," << random_str << endl;
         generated++;
         attempts++;
